requirements.txt
fastapi
uvicorn
python-dotenv
requests
urllib3
pydantic
langchain
langchain-community
langchain-openai
langchain-core
pymupdf
scikit-learn
numpy
rank_bm25
faiss-cpu

# Multi-format document processing dependencies
python-docx>=1.1.0
python-pptx>=1.0.0
openpyxl>=3.1.0
pillow>=11.0.0
pytesseract>=0.3.10
lxml>=5.2.0
unstructured[all-docs]>=0.12.0
python-magic>=0.4.27

